signup refactor
