+++ client/src/interfaces/interfaces.ts
@@ -2,14 +2,6 @@ export interface LoggedInProps {
   isLoggedIn: boolean;
 }
 
-export interface SignupFields {
-  username: string;
-  password: string;
-  confirmPassword: string;
-  securityQuestion: string;
-  securityAnswer: string;
-}
-
 export interface SignupVals {
   username: string;
   password: string;
